refactor to gorilla mux
