+++ cmd/api/healthcheck.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	"fmt"
+	"net/http"
+)
+
+func (srv *application) healthcheckHandler(w http.ResponseWriter, r *http.Request) {
+	fmt.Fprintln(w, "status: available")
+}
